// src/constants/bot-messages-keys.enum.ts
/**
 * Keys for bot messages from localization files.
 */
export enum BotMessagesKeys {
    USERNAME_IS_EVEN_WITH_EVERYONE = 'USERNAME_IS_EVEN_WITH_EVERYONE',
    USERNAME_SHOULD_PAY_TO_GROUP = 'USERNAME_SHOULD_PAY_TO_GROUP',
    USERNAME_GETS_CASHBACK_FROM_GROUP = 'USERNAME_GETS_CASHBACK_FROM_GROUP',
    ITEMS_SPEND_MONEY_ON = 'ITEMS_SPEND_MONEY_ON',
    COUNT_MONEY_RESPONSE = 'COUNT_MONEY_RESPONSE',
    COUNT_MONEY_RESULTS = 'COUNT_MONEY_RESULTS',
    NOTHING_POSTED_IN_CHAT_YET = 'NOTHING_POSTED_IN_CHAT_YET',
}